media: i2c: Fix array crop registers in ov2740
The ov2740 driver supports a single mode which sets 1932x1092 output.
The driver defines a crop rectangle that is larger than that, and it
also starts the rectangle off-centred at 0,0 offset. Change the crop
registers so that the crop rectangle matches the output for the mode
and is taken from the centre of the pixel array.

The settings are also duplicated for some reason - remove the dupes.

Signed-off-by: Daniel Scally <djrscally@gmail.com>